requirements.txt
# abstract_singleton==1.0.1
# aiohttp==3.8.4
# aiosignal==1.3.1
# anyio==3.6.2
# async-generator==1.10
# async-timeout==4.0.2
# asynctest==0.13.0
# attrs==23.1.0
# auto_gpt_plugin_template==0.0.3
# autoflake==2.1.1
# beautifulsoup4==4.12.2
# black==23.3.0
# blinker==1.8.2
# blis==0.7.9
# cachetools==5.3.0
# catalogue==2.0.8
# certifi==2022.12.7
# cffi==1.15.1
# cfgv==3.3.1
# chardet==5.1.0
# charset-normalizer==3.1.0
# click==8.1.3
# colorama==0.4.6
# confection==0.0.4
# contourpy==1.2.0
# coverage==7.2.5
# cryptography==40.0.2
# cssselect==1.2.0
# cycler==0.12.1
# cymem==2.0.7
# diskcache==5.6.1
# distlib==0.3.6
# distro==1.8.0
# dnspython==2.3.0
# docker==6.1.0
# docstring_parser==0.16
# duckduckgo-search==2.9.3
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
# exceptiongroup==1.1.1
# filelock==3.12.0
# flake8==6.0.0
# Flask==3.0.3
# Flask-Cors==5.0.0
# Flask-JWT-Extended==4.6.0
# Flask-SQLAlchemy==3.1.1
# fonttools==4.49.0
# frozenlist==1.3.3
# ghp-import==2.1.0
# gitdb==4.0.10
# GitPython==3.1.31
# google-api-core==2.11.0
# google-api-python-client==2.86.0
# google-auth==2.34.0
# google-auth-httplib2==0.1.0
# google-cloud-aiplatform==1.65.0
# google-cloud-bigquery==3.25.0
# google-cloud-core==2.4.1
# google-cloud-resource-manager==1.12.5
# google-cloud-storage==2.14.0
# google-crc32c==1.6.0
# google-resumable-media==2.7.2
# googleapis-common-protos==1.59.0
# greenlet==3.1.0
# grpc-google-iam-v1==0.13.1
# grpcio==1.66.1
# grpcio-status==1.62.3
# gTTS==2.3.1
# h11==0.14.0
# httpcore==0.17.0
# httplib2==0.22.0
# httpx==0.24.0
# identify==2.5.24
# idna==3.4
# iniconfig==2.0.0
# isort==5.12.0
# itsdangerous==2.2.0
# Jinja2==3.1.2
# jsonschema==4.17.3
# kiwisolver==1.4.5
# langcodes==3.3.0
# loguru==0.7.0
# lxml==4.9.2
# Markdown==3.3.7
# MarkupSafe==2.1.2
# matplotlib==3.8.3
# mccabe==0.7.0
# mergedeep==1.3.4
# mkdocs==1.4.3
# multidict==6.0.4
# murmurhash==1.0.9
# mypy-extensions==1.0.0
# mysql-connector-python==9.0.0
# nodeenv==1.7.0
# numpy==1.24.3
# oauthlib==3.2.2
# openai==0.27.2
# openapi-python-client==0.13.4
# orjson==3.8.10
# outcome==1.2.0
# packaging==23.1
# pathspec==0.11.1
# pathy==0.10.1
# Pillow==9.5.0
# pinecone-client==2.2.1
# platformdirs==3.5.0
# playsound==1.2.2
# pluggy==1.0.0
# pre-commit==3.3.1
# preshed==3.0.8
# proto-plus==1.24.0
# protobuf==4.22.4
# psycopg2==2.9.9
# psycopg2-binary==2.9.9
# py-cpuinfo==9.0.0
# pyasn1==0.5.0
# pyasn1-modules==0.3.0
# pycodestyle==2.10.0
# pycparser==2.21
# pydantic==1.10.7
# pyflakes==3.0.1
# PyJWT==2.9.0
# pymdown-extensions==9.11
# pyOpenSSL==23.1.1
# pyparsing==3.0.9
# pyrsistent==0.19.3
# PySocks==1.7.1
# pytest==7.3.1
# pytest-asyncio==0.21.0
# pytest-benchmark==4.0.0
# pytest-cov==4.0.0
# pytest-integration==0.2.3
# pytest-mock==3.10.0
# pytest-recording==0.12.2
# python-dateutil==2.8.2
# python-dotenv==1.0.0
# PyYAML==6.0
# pyyaml_env_tag==0.1
# readability-lxml==0.8.1
# redis==4.5.4
# regex==2023.5.5
# requests==2.30.0
# requests-oauthlib==1.3.1
# rsa==4.9
# selenium==4.1.4
# shapely==2.0.6
# shellingham==1.5.0.post1
# six==1.16.0
# smart-open==6.3.0
# smmap==5.0.0
# sniffio==1.3.0
# sortedcontainers==2.4.0
# soupsieve==2.4.1
# spacy==3.5.2
# spacy-legacy==3.0.12
# spacy-loggers==1.0.4
# SQLAlchemy==2.0.34
# srsly==2.4.6
# thinc==8.1.10
# tiktoken==0.3.3
# tqdm==4.65.0
# trio==0.22.0
# trio-websocket==0.10.2
# ttkbootstrap==1.10.1
# tweepy==4.14.0
# typer==0.7.0
# typing_extensions==4.12.2
# uritemplate==4.1.1
# urllib3==1.26.15
# urllib3-secure-extra==0.1.0
# vcrpy==4.2.1
# virtualenv==20.23.0
# wasabi==1.1.1
# watchdog==3.0.0
# webdriver-manager==3.8.6
# websocket-client==1.5.1
# Werkzeug==3.0.4
# wrapt==1.15.0
# wsproto==1.2.0
# yarl==1.9.2


Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Migrate==4.0.5
Werkzeug==2.3.8
psycopg2-binary==2.9.9
python-dotenv==1.0.0
SQLAlchemy==2.0.34
Jinja2==3.1.2
click==8.1.3
MarkupSafe==2.1.5
email-validator==2.1.0.post1
Pillow==9.5.0